feat: rename component as Input With Checkboxes (Issue #28)
