+++ src/components/Input/InputWithCheckboxes/InputWithCheckboxes.styled.ts
